Add the new nick models to channel/server and set up the view
I renamed the ChannelNickListProxyModel to
ChannelNickListFilterModel because it does more filtering than
proxying. Then I put the models in channel/server and fixed
a couple of crashes that this exposed. Then I created a
QListView that lives right below the old nicklistview and will
allow me to view the state of both nick lists side by side
so I can know when I've got everything perfect.