Connect model with join/leave/nick setting changes
There are now currently 3 nicklistviews shown: the old one, the
new channel nick list, and the nicklistmodel so that I can
monitor that the nicks are being deleted from the whole model
when they should be. Currently the nicks aren't properly deleted
when you leave a channel, I will have to iterate over all of the
nicks in order to do that, I'm trying to think of a better way.

Numerous bugs fixed, SizeHintRole works now, icons work, away
works correctly, and setting modes for nicks works now.