Use one nicklistmodel for all servers
This will make it much easier to do the nicksonline page and any
other things in the future that will require access to all nicks
on all servers.

While I was at it I also switched around the arguments in many of
the functions so it goes in descending order of specificity
e.g. server->channel->nick->property.