Convert serverlistdialog to a model/view
I have done this by creating a servergroupmodel which I will use
for the NicksOnlineModel too. This still requires testing and
has a few bugs and missing features.

The notifyLists were moved inside the servergroups both to match
how they are in the config, and to make it easier to use the
planned servergroupmodel/nicksonlinemodel combo.

Due to that move much of the nicksonline functionality has been
temporarily disabled. It will be replaced in future commits.