Initial commit of the new nicksonline
Completely rewritten to use the new servergroupmodel with
the beginnings of support for MONITOR and WATCH.

Although it sends the ISON commands currently they don't do
anything, and there are some other bugs that prevent it from
working as expected. This is only an initial commit so I'll
fix those as I go along.