Fix numerous drag/drop problems with the children of groups
The servers (and in theory notify nicks) now drag and drop
as expected except in the case that they are dropped on the last
row of a different server group.

All crash bugs appear to be fixed by checking beginMoveRows
and returning false if it says the move is invalid.