Fix a bug where invalid rows were shown in server group views
Rows were shown even if there was nothing to show for that
column. E.g. if there were 4 notify nicks and 3 servers
in a server group, the server list dialog would have 1 empty
row at the bottom of its' list. This was fixed by filtering
the rows and getting a row count for that column.

Nicksonline can probably filter empty servergroups too
and I believe that is old behavior, but I'm leaving it for now
to debug some other issues.

Added a TODO to emit changes when a delete occurs in slotDelete