Fix a bug where the second child column wasn't showing
Instead of picking which columns to show now we just take the
first two and take the information we want from the column we
want. e.g. we show the server column, but get the displayrole
from the notify nick source column.

Also placed some todo's for changing the selection when adding
and removing nicks.