Fix a bug where the rowcount was off for notify nicks
A bit of a logic error here, the source model should have a row
for every column if one of the columns has data, because a tree
can't have uneven rows. Now the rowcount in the servergroupmodel
is the qMax of all the rowcounts, and the filter models use their
own counts to filter out the rows they don't need.