Finally fix the online status of watched nicks.
Before we were only getting the online status of unwatched nicks.
We only watch nicks if we share no channels with them, if we do
share a channel it's easy to tell that they're online.

Now we're properly using ISON again, and the timer has been
adjusted to only run at the interval after the response.
Meaning if we don't get the right response from the server we
won't keep spamming it with ISON's.