Replace watched nick with blank nick after it goes offline
When a nick goes offline, and we've made sure by adding them
to a nick-watching service, we need to remove the current nick
object, because it could contain hostmask/netserver info that is
now invalid, and would be displayed if they logged in again. To
avoid this we create a brand new nick object, and replace the
current on in m_watchedNicks.