Modify nicks in both nicklistmodel/nicksonline model and whois
The inputfilter will now try to modify nicks in the nicklistmodel
and the nicksonline model depending on which is online and if
the user wants to useNotify().

There is also now support in nicksonline to send out whois
for each of the nicks that are online, when they come online
with a list to make sure we only send it out once per nick
at a time.

Other fixes to server/channel so they use the models properly