Use nicks to affect the model instead of vice versa
Before the model was the gobetween for the nick and would use
that position to signal if data was changed in the model.
Now the model merely supplies the nick, and then has a slot
so that it is alerted when the nick changes, and the nick signals
whenever it is changed.

Then I changed it so that instead of trying to get the nick
to find out information, the proxy model goes through the
model for all of it's data. I may be able to do away with
the internal pointer entirely.

I also fixed the last sorting bug, I had mixed up sorting order
with some 'flags' that were used to tell if the nicklist needed
to be updated.