Initial commit for MONITOR/WATCH support
This commit split up the watched nicks into 'watched' and
'actively watched.' An actively watched nick is one we use
monitor/watch/ison for to determine if they are online. If they
are not in the activelywatched list, then they are passively
watched either because we are in the same channel, or because
(after I add the check in) the user has that feature disabled.

All monitor/watch/ison commands are now buffered and inputfilter
should only send the output to the user if they requested it
themselves.

Some of the i18n strings aren't very good, so they'll probably
need to be replaced with something else.